Extract seperate files for dealer types in service
