+++ Domain Model/EurofurenceModel/Public/Objects/Entities/Dealer/Dealer.swift
